Torque32bits

+++ eth/embobj/plus/comm-v2/icub/EoMeasures.h
@@ -102,14 +102,14 @@ typedef int32_t             eOmeas_acceleration_t;
     @brief      eOmeas_force_t contains the force values.
                 the measurement unit is mN (milli Newton).
  **/
-typedef int16_t             eOmeas_force_t;
+typedef int32_t             eOmeas_force_t;
 
 
-/** @typedef    typedef int16_t             eOmeas_torque_t
+/** @typedef    typedef int32_t             eOmeas_torque_t
     @brief      eOmeas_torque_t contains the torque values.
                 the measurement unit is mN m (milli Newton metre).
  **/
-typedef int16_t             eOmeas_torque_t;
+typedef int32_t             eOmeas_torque_t;
 
 
 /** @typedef    typedef int16_t             eOmeas_current_t
